refactor: pygen warnings are now handleg internally
