+++ cognite/pygen/_generator.py
@@ -20,10 +20,10 @@ from cognite.pygen._core.generators import SDKGenerator
 from cognite.pygen._core.models import DataClass
 from cognite.pygen._settings import _load_pyproject_toml
 from cognite.pygen._version import __version__
+from cognite.pygen._warnings import InvalidCodeGenerated, print_warnings
 from cognite.pygen.config import PygenConfig
 from cognite.pygen.exceptions import DataModelNotFound
 from cognite.pygen.utils.text import to_pascal, to_snake
-from cognite.pygen.warnings import InvalidCodeGenerated, print_warnings
 
 DataModel = Union[DataModelIdentifier, dm.DataModel[dm.View]]
