+++ cognite/pygen/_warnings.py
